Unify AssetWaitTool
