+++ src/tools/asset-generate/cinematic/cinematic-generation.ts
@@ -423,95 +423,3 @@ Use this tool to check the current status of a cinematic generation job in the q
     }
   }
 }
-
-/**
- * Cinematic Wait Tool
- *
- * A utility tool that allows the LLM to wait a specified amount of time before checking status again.
- */
-export class CinematicWaitTool {
-  name = 'cinematic_wait';
-  description = `Allows waiting for a specified number of seconds before continuing execution.
-
-Use this tool between status checks to give time for cinematic generation to progress.
-The recommended wait time is 30 seconds between status checks.`;
-
-  inputSchema = {
-    type: 'object',
-    properties: {
-      seconds: {
-        type: 'integer',
-        description: 'Number of seconds to wait. Default is 30 seconds.',
-        default: 30,
-        minimum: 1,
-        maximum: 120,
-      },
-      status_message: {
-        type: 'string',
-        description: 'Optional status message to display during the wait',
-      },
-    },
-    required: ['seconds'],
-  };
-
-  async execute(args: Record<string, any>, context: ToolExecutionContext): Promise<any> {
-    try {
-      const seconds = args.seconds || 30;
-      const statusMessage =
-        args.status_message || `Waiting ${seconds} seconds for cinematic generation...`;
-
-      // Limit waiting time to a reasonable amount (between 1 and 120 seconds)
-      const waitTime = Math.max(1, Math.min(120, seconds));
-
-      // Initial progress update
-      if (context.progressCallback) {
-        await context.progressCallback({
-          progress: 0,
-          total: waitTime,
-          message: statusMessage,
-        });
-      }
-
-      // Wait loop with progress updates
-      for (let i = 1; i <= waitTime; i++) {
-        // Wait for 1 second
-        await new Promise((resolve) => setTimeout(resolve, 1000));
-
-        // Update progress every second
-        if (context.progressCallback) {
-          await context.progressCallback({
-            progress: i,
-            total: waitTime,
-            message: `${statusMessage} (${i}/${waitTime}s)`,
-          });
-        }
-      }
-
-      // Final progress update
-      if (context.progressCallback) {
-        await context.progressCallback({
-          progress: waitTime,
-          total: waitTime,
-          message: 'Wait completed. You can now check the status again.',
-        });
-      }
-
-      return {
-        content: [
-          {
-            type: 'text',
-            text: `Waited for ${waitTime} seconds. You can now check the status of your cinematic generation.`,
-          },
-        ],
-        isError: false,
-      };
-    } catch (error: unknown) {
-      const errorMessage = error instanceof Error ? error.message : String(error);
-      logger.error('Failed to execute wait:', error);
-      return {
-        content: [{ type: 'text', text: `Error during wait: ${errorMessage}` }],
-        isError: true,
-      };
-    }
-  }
-}
