+++ src/tools/asset-generate/index.ts
@@ -3,21 +3,19 @@ import {
   CinematicAssetGeneratorTool,
   CinematicAssetResultTool,
   CinematicStatusTool,
-  CinematicWaitTool,
 } from './cinematic/cinematic-generation.js';
 import {
   MusicGeneratorTool,
   SoundEffectsGeneratorTool,
   AudioStatusTool,
   AudioResultTool,
-  AudioWaitTool,
 } from './audio/index.js';
 import {
   SkyboxGeneratorTool,
   SkyboxStatusTool,
-  SkyboxWaitTool,
   SkyboxStylesTool,
 } from './skybox/skybox-generation.js';
+import { AssetWaitTool } from './common/wait-tool.js';
 
 // Comprehensive asset generation tools collection
 export const assetGenerateTools = [
@@ -28,20 +26,20 @@ export const assetGenerateTools = [
   new CinematicAssetGeneratorTool(),
   new CinematicAssetResultTool(),
   new CinematicStatusTool(),
-  new CinematicWaitTool(),
 
   // Audio generation tools
   new MusicGeneratorTool(),
   new SoundEffectsGeneratorTool(),
   new AudioStatusTool(),
   new AudioResultTool(),
-  new AudioWaitTool(),
 
   // Skybox generation tools
   new SkyboxGeneratorTool(),
   new SkyboxStatusTool(),
-  new SkyboxWaitTool(),
   new SkyboxStylesTool(),
+
+  // Shared tools
+  new AssetWaitTool(),
 ];
 
 // Export individual tools
@@ -53,18 +51,18 @@ export {
   CinematicAssetGeneratorTool,
   CinematicAssetResultTool,
   CinematicStatusTool,
-  CinematicWaitTool,
 
   // Audio related
   MusicGeneratorTool,
   SoundEffectsGeneratorTool,
   AudioStatusTool,
   AudioResultTool,
-  AudioWaitTool,
 
   // Skybox related
   SkyboxGeneratorTool,
   SkyboxStatusTool,
-  SkyboxWaitTool,
   SkyboxStylesTool,
+
+  // Shared tools
+  AssetWaitTool,
 };
